move some utilities out of python completer
